DB/pyproject.toml
[project]
name = "db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pymongo>=4.13.2",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.35.0",
]
